refactor: components/checkout -> components/Checkout
