+++ packages/commercetools/theme/pages/Checkout.vue
@@ -21,8 +21,8 @@
 <script>
 
 import { SfSteps, SfButton } from '@storefront-ui/vue';
-import CartPreview from '~/components/checkout/CartPreview';
-import OrderReview from '~/components/checkout/OrderReview';
+import CartPreview from '~/components/Checkout/CartPreview';
+import OrderReview from '~/components/Checkout/OrderReview';
 import { ref, computed } from '@vue/composition-api';
 import { useUser } from '@vue-storefront/commercetools';
