+++ packages/core/nuxt-theme-module/theme/components/Checkout/CartPreview.vue
