+++ packages/core/nuxt-theme-module/theme/pages/Checkout.vue
@@ -25,8 +25,8 @@
 <script>
 
 import { SfSteps, SfButton } from '@storefront-ui/vue';
-import CartPreview from '~/components/checkout/CartPreview';
-import OrderReview from '~/components/checkout/OrderReview';
+import CartPreview from '~/components/Checkout/CartPreview';
+import OrderReview from '~/components/Checkout/OrderReview';
 import { ref } from '@vue/composition-api';
 
 const STEPS = [
